+++ src/main/scala/me/piotrjander/cinema/application/provider/ConfirmationSecretUUIDGenerator.scala
@@ -1,4 +1,4 @@
-package me.piotrjander.cinema.application
+package me.piotrjander.cinema.application.provider
 
 import java.util.UUID
