+++ src/main/scala/me/piotrjander/cinema/application/service/ReservationService.scala
@@ -3,9 +3,10 @@ package me.piotrjander.cinema.application.service
 import cats.ApplicativeError
 import cats.effect.Async
 import cats.implicits._
-import me.piotrjander.cinema.application.{ConfirmationSecretGenerator, EntityPayloads, LocalClock}
+import me.piotrjander.cinema.application.EntityPayloads
 import me.piotrjander.cinema.application.message.ReservationMessage._
 import me.piotrjander.cinema.application.exception.BadRequestException
+import me.piotrjander.cinema.application.provider.{ConfirmationSecretGenerator, LocalClock}
 import me.piotrjander.cinema.application.validator.{FullNameValidator, TicketsBreakdownValidator}
 import me.piotrjander.cinema.domain.entity.{Reservation, ReservationRequest, ScreeningId}
 import me.piotrjander.cinema.domain.repository._
